審計+硬化：LLBC ingest 轉真 serde_json（frontends/llbc）+ 鏟除假 Groth16 prover + 死代碼/警告清零 + clippy 機械修復
- C0–C4 接口鏈複檢：真 rustc/charon → LLBC → IR → 出口，全鏈無 mock/stub/fake
- core 手寫 JSON lexer 刪除（~190 行）；文本→樹 = frontends/llbc 真 serde_json；
  core 保留零依賴 Value 樹 + 法證白名單 schema walker（LlbcRoot::from_value）；
  core 測試經 serde_json（dev-dep only，[dependencies] 繼續為空）
- 鏟除 QAP「Groth16 prover」假實現（DefaultHasher 假 proof + verified:true 硬編碼 + 假 on-chain verifier）＋
  lib.rs 描述誠實修正＋PHASE_B 過度宣稱記入審計報告
- 死賦值清零（rustc 證明下准去 init / attempts 等式化）；git rm 意外入倉 main.rs.bak
- cargo clippy --fix 全 workspace（51 檔，僅 MachineApplicable 等價改寫）+ IR 骨架 dead_code 保真聲明
- lib warnings 24 → 0；全量測試 126+4 全綠

+++ core/src/solana_onchain.rs
@@ -645,8 +645,8 @@ pub fn payload_from_v3_result(v3: &crate::pipeline_v3::PipelineV3Result) -> Sola
     let last = v3.iterations.last();
     SolanaQapPayload {
         source_name: v3.source_name.clone(),
-        r1cs_constraints: last.map(|it| it.n_vars).unwrap_or(100) as usize, // 簡化用 n_vars
-        r1cs_wires: last.map(|it| it.n_polys).unwrap_or(50) as usize,
+        r1cs_constraints: last.map(|it| it.n_vars).unwrap_or(100), // 簡化用 n_vars
+        r1cs_wires: last.map(|it| it.n_polys).unwrap_or(50),
         risk_score: v3.commercial.risk_score,
         qap_verified: last.and_then(|it| it.qap_verified).unwrap_or(false),
         qap_tamper_rejected: last.and_then(|it| it.qap_tamper_rejected).unwrap_or(false),
